+++ src/lib/libdpd/init.c
@@ -16,7 +16,7 @@ typedef struct {
     int ler;
 } dpdpair;
 
-int dpd_init(int dpd_num, int nirreps, int memory, int cachetype,
+int dpd_init(int dpd_num, int nirreps, long int memory, int cachetype,
              int *cachefiles, int **cachelist, 
              struct dpd_file4_cache_entry *priority, int num_subspaces, ...)
 {
